@@ -21,14 +21,22 @@ package main_test
 
 import (
 	"context"
-	"io/ioutil"
+	"net"
 	"os"
 	"path/filepath"
 	"testing"
 	"time"
 
 	"github.com/edwarnicke/exechelper"
+	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
+
+	"github.com/networkservicemesh/api/pkg/api/networkservice"
+	"github.com/networkservicemesh/api/pkg/api/networkservice/mechanisms/cls"
+	"github.com/networkservicemesh/api/pkg/api/registry"
+	"github.com/networkservicemesh/nsmctl/internal/pkg/tools/domain"
+	"github.com/networkservicemesh/nsmctl/internal/pkg/tools/persistence"
+	"github.com/networkservicemesh/sdk/pkg/tools/sandbox"
 )
 
 type MainSuite struct {
@@ -64,7 +72,7 @@ func (s *MainSuite) Test_Generate_NetworkServiceEndpoint() {
 
 	s.RequireExec("nsmctl gen nse --name nse-1 --labels app=my-nse,version=v1.0.0 --path " + dir)
 
-	files, err := ioutil.ReadDir(dir)
+	files, err := os.ReadDir(dir)
 
 	s.Require().NoError(err)
 
@@ -82,13 +90,80 @@ func (s *MainSuite) Test_Generate_NetworkServiceEndpointVpp() {
 
 	s.RequireExec("nsmctl gen nse vpp --name nse-1 --labels app=my-nse,version=v1.0.0 --path " + dir)
 
-	files, err := ioutil.ReadDir(dir)
+	files, err := os.ReadDir(dir)
 
 	s.Require().NoError(err)
 
 	s.Require().Len(files, 6)
 
 	s.RequireExec("go build ./...", exechelper.WithDir(dir))
+}
+
+func (s *MainSuite) Test_SandboxAndNSMControl() {
+	var ctx, cancel = context.WithCancel(s.ctx)
+	defer cancel()
+	var d = sandbox.NewBuilder(ctx, s.T()).SetNodesCount(1).Build()
+
+	defer func() {
+		_ = persistence.Delete[*domain.Domain]("test")
+	}()
+
+	_ = persistence.Store("test", &domain.Domain{
+		Name:            "test",
+		ManagerService:  net.JoinHostPort(d.Nodes[0].NSMgr.URL.Hostname(), d.Nodes[0].NSMgr.URL.Port()),
+		RegistryService: net.JoinHostPort(d.Registry.URL.Hostname(), d.Registry.URL.Port()),
+		IsInsecure:      true,
+	})
+
+	nseReg := &registry.NetworkServiceEndpoint{
+		Name:                "final-endpoint",
+		NetworkServiceNames: []string{"ns"},
+	}
+
+	nsRegistryClient := d.NewNSRegistryClient(ctx, sandbox.GenerateTestToken)
+
+	_, err := nsRegistryClient.Register(ctx, &registry.NetworkService{Name: "ns"})
+	require.NoError(s.T(), err)
+
+	_ = d.Nodes[0].NewEndpoint(ctx, nseReg, sandbox.GenerateTestToken)
+
+	nsc := d.Nodes[0].NewClient(ctx, sandbox.GenerateTestToken)
+
+	request := &networkservice.NetworkServiceRequest{
+		MechanismPreferences: []*networkservice.Mechanism{
+			{Cls: cls.LOCAL, Type: "kernel"},
+		},
+		Connection: &networkservice.Connection{
+			Id:             "1",
+			NetworkService: "ns",
+			Labels:         make(map[string]string),
+		},
+	}
+
+	_, err = nsc.Request(ctx, request)
+	require.NoError(s.T(), err)
+
+	s.RequireExec("nsmctl get domains --domain test")
+	s.RequireExec("nsmctl get nses --domain test")
+	s.RequireExec("nsmctl get netsvc --domain test")
+	s.RequireExec("nsmctl get connections --domain test")
+
+	s.RequireExec("nsmctl describe domains")
+	s.RequireExec("nsmctl describe nses --domain test")
+	s.RequireExec("nsmctl describe netsvc --domain test")
+	s.RequireExec("nsmctl describe connections --domain test")
+
+	var p = filepath.Join(s.T().TempDir(), "mse.yaml")
+	_ = os.WriteFile(p, []byte("name: my-nse"), os.ModePerm)
+	s.RequireExec("nsmctl apply nse --domain test -f " + p)
+	s.RequireExec("nsmctl get nse --domain test my-nse")
+	s.RequireExec("nsmctl delete nse --domain test my-nse")
+
+	p = filepath.Join(s.T().TempDir(), "ns.yaml")
+	_ = os.WriteFile(p, []byte("name: my-ns"), os.ModePerm)
+	s.RequireExec("nsmctl apply netsvc --domain test -f " + p)
+	s.RequireExec("nsmctl get netsvc --domain test my-ns")
+	s.RequireExec("nsmctl delete netsvc --domain test my-ns")
 }
 
 func Test_RunSystemTests(t *testing.T) {
